feat: Created pokemon.js file to organize all api requests for pokemons instead of two diferent files
