--- src/api.js
@@ -1,5 +1,0 @@
-import axios from 'axios';
-
-export const api = axios.create({
-  baseURL: 'https://pokeapi.co/api/v2',
-})
